Fix: SafeArray of immidiate can't access.
SafeArrayGetElement return immidate value when safe array is array of UI1.
So if you want to return value to VARIANT, you should set V_VT for input val.
Just setting this value must be enough for this fix, but I also change to use SafeArrayPtrOfIndex.
Because ruby version of win32ole use similar code, and it's used much wider and so it must be much robust.
(And it's a little faster, anyway).
Reference: http://d.hatena.ne.jp/miyamuko/20060619/p1