update documentation for behaviour change in `eval-in-another-xyzzy`
In old implementation, which is a macro you can write BODY forms like
`progn` and evaluate it in another xyzzy, it is not possible to refer
variable bindings in current environment. This is different from what
you expect and need to be noted.

In new implementation, wihch is a function you give a FORM like `eval`
and evaluate it in another xyzzy, this is still impossible.
But since new implementation is analogue to `eval`, this behaviour is
what you expect.

Actually, it now allows you to "implant" variables by using backquoted
template:

    (let ((name "John Irenicus"))
      (eval-in-another-xyzzy
       `(format nil "Hello, ~A" ,name)))
    => "Hello, John Irenicus"

It is still subject to restriction that all elements of FORM must be
an object that can be printed in reable representation, and they are
not same object if referred by comma more than once, so you need to
be bit careful to do this.